requirements.txt
import tkinter as tk
import PyPDF2 
from PIL import Image, ImageTk
Pip installations

asgiref                  3.6.0
certifi                  2022.12.7
charset-normalizer       3.1.0
Django                   4.2
duckdb                   0.7.1
fpdf                     1.7.2
gcloud                   0.18.3
googleapis-common-protos 1.59.0
httplib2                 0.22.0
idna                     3.4
image                    1.5.33
numpy                    1.24.3
oauth2client             4.1.3
pandas                   2.0.1
Pillow                   9.5.0
protobuf                 4.22.3
pyarrow                  11.0.0
pyasn1                   0.5.0
pyasn1-modules           0.3.0
pylance                  0.4.4
pyparsing                3.0.9
PyPDF2                   3.0.1
python-dateutil          2.8.2
pytz                     2023.3
requests                 2.28.2
rsa                      4.9
six                      1.16.0
sqlparse                 0.4.4
tzdata                   2023.3
urllib3                  1.26.15